refactor(appserver): isolate agent thread sync
